Integrated the datamanager module with the diagnosenet library

+++ diagnosenet/optimizers.py
@@ -8,17 +8,15 @@ import tensorflow as tf
 class Optimizer:
     def __init__(self) -> None:
         self.lr: float
-        
+
     def desktop_Grad(self, loss) -> tf.Tensor:
         raise NotImplementedError
-        
-        
+
+
 class Adam(Optimizer):
-    def __init__(self, lr: int = 0.001) -> None:
+    def __init__(self, lr: float = 0.001) -> None:
         super().__init__()
         self.lr = lr
-        
+
     def desktop_Grad(self, loss) -> tf.Tensor:
         return tf.train.AdamOptimizer(self.lr).minimize(loss)
-        
-
